sql updated, some jni updated. sql builds without errors.

+++ Planner-Gui/app/src/main/cpp/sqlite-dao.c
@@ -25,7 +25,7 @@ bool createTableQuery(const char *filepath) {
 
     const char *createWeeklyQuery = "CREATE TABLE IF NOT EXISTS WEEKLY(weeklyID AUTOINCREMENT NOT NULL, monday Boolean NOT NULL,"
             "tuesday Boolean NOT NULL,"
-            "wenesday Boolean NOT NULL,"
+            "wednesday Boolean NOT NULL,"
             "thursday Boolean NOT NULL,"
             "friday Boolean NOT NULL,"
             "saturday Boolean NOT NULL,"
@@ -34,13 +34,17 @@ bool createTableQuery(const char *filepath) {
 
     const char *createCALENDARQuery = "CREATE TABLE IF NOT EXISTS CALENDAR("
             "eventID AUTOINCREMENT NOT NULL,"
-            "date DATE NOT NULL,"
+            "day text NOT NULL,"
+            "month text NOT NULL,"
+            "year text NOT NULL,"
             "title text NOT NULL,"
             "description text NOT NULL,"
-            "start time NOT NULL,"
-            "duration int NOT NULL, "
+            "start text NOT NULL,"
+            "duration text NOT NULL, "
             "repeatCycle int NOT NULL,"
-            "endDate date NOT NULL,"
+            "endDay text NOT NULL,"
+            "endMonth text NOT NULL,"
+            "endYear text NOT NULL,"
             "weeklyID NOT NULL, "
             "PRIMARY KEY (eventID)),"
             "FOREIGN KEY(weeklyID) REFERENCES WEEKLY(weeklyID)) ON DELETE CASCADE;";
@@ -53,36 +57,52 @@ bool createTableQuery(const char *filepath) {
     return true;
 }
 
-bool createDb(const char *createStatment, const char *filepath) {
+bool createDb(const char *createStatement, const char *filepath) {
     /* Open database */
     sqlite3 *db;
     char *zErrMsg = 0;
     int rc;
     rc = sqlite3_open(filepath, &db);
 
-    rc = sqlite3_exec(db, createStatment, callback, 0, &zErrMsg);
+
     if (rc != SQLITE_OK) {
         sqlite3_free(zErrMsg);
         return false;
     }
+    rc = sqlite3_exec(db, createStatement, callback, 0, &zErrMsg);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
+
     sqlite3_close(db);
     return true;
 }
 
-bool insertToDb(const char *date, const char *title, const char *description, const char *start,
-                const char *duration,
-                int repeatCycle, const char *endDate, bool repeats[7], const char *filepath) {
+bool insertToDb(const char *day, const char *month, const char *year, const char *title,
+                const char *description, const char *start, const char *duration,
+                int repeatCycle, const char *endDay, const char *endMonth, const char *endYear,
+                bool repeats[7], const char *filepath) {
+
+
     sqlite3 *db;
     char *zErrMsg = 0;
-    int rc, i, j;
-    int repeater = repeats;
+    int rc, i;
+    //int repeater = repeats;
     sqlite3_stmt *stmt;
 
     /* Open database */
     rc = sqlite3_open(filepath, &db);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
 
-    char *insertQueryWeekly = "INSERT INTO WEEKLY(monday,tuesday,wednesday,thursday,friday,saturday,sunday) VALUES(?,?,?,?,?,?,?);";
-    char *insertQueryCalendar = "INSERT INTO CALENDAR(date,title,description,start,duration,repeatCycle,endDate,weeklyID) VALUES(?,?,?,?,?,?,?,?,?,last_insert_rowid());";
+    char *insertQueryWeekly = "INSERT INTO WEEKLY(monday,tuesday,wednesday,thursday,friday,saturday,"
+                                "sunday) VALUES(?,?,?,?,?,?,?);";
+    char *insertQueryCalendar = "INSERT INTO CALENDAR(day,month,year,title,description,start,"
+                                "duration,repeatCycle,endDay,endMonth,endYear,weeklyID) "
+                                "VALUES(?,?,?,?,?,?,?,?,?,?,?,last_insert_rowid());";
 
     printf("Query %s\n", insertQueryWeekly);
     printf("Query %s\n", insertQueryCalendar);
@@ -91,7 +111,7 @@ bool insertToDb(const char *date, const char *title, const char *description, co
     if (rc == SQLITE_OK) {
 
         for (i = 0; i < 7; i++) {
-            int days = sqlite3_bind_int(stmt, i + 1, repeats[i]);
+            sqlite3_bind_int(stmt, i + 1, repeats[i]);
         }
 
     } else {
@@ -103,13 +123,17 @@ bool insertToDb(const char *date, const char *title, const char *description, co
 
     rc = sqlite3_prepare_v2(db, insertQueryCalendar, -1, &stmt, 0);
     if (rc == SQLITE_OK) {
-        sqlite3_bind_text(stmt, 1, date, strlen(date), 0);
-        sqlite3_bind_text(stmt, 2, title, strlen(title), 0);
-        sqlite3_bind_text(stmt, 3, description, strlen(description), 0);
-        sqlite3_bind_text(stmt, 4, start, strlen(start), 0);
-        sqlite3_bind_text(stmt, 5, duration, strlen(duration), 0);
-        sqlite3_bind_text(stmt, 6, repeatCycle, strlen(repeatCycle), 0);
-        sqlite3_bind_text(stmt, 7, endDate, strlen(endDate), 0);
+        sqlite3_bind_text(stmt, 1, day, (int)strlen(day), 0);
+        sqlite3_bind_text(stmt, 2, month, (int)strlen(month), 0);
+        sqlite3_bind_text(stmt, 3, year, (int)strlen(year), 0);
+        sqlite3_bind_text(stmt, 4, title, (int)strlen(title), 0);
+        sqlite3_bind_text(stmt, 5, description, (int)strlen(description), 0);
+        sqlite3_bind_text(stmt, 6, start, (int)strlen(start), 0);
+        sqlite3_bind_text(stmt, 7, duration, (int)strlen(duration), 0);
+        sqlite3_bind_int(stmt, 8, repeatCycle);
+        sqlite3_bind_text(stmt, 9, endDay, (int)strlen(endDay), 0);
+        sqlite3_bind_text(stmt, 10, endMonth, (int)strlen(endMonth), 0);
+        sqlite3_bind_text(stmt, 11, endYear, (int)strlen(endYear), 0);
     } else {
         fprintf(stderr, "Failed to execute statement: %s\n", sqlite3_errmsg(db));
     }
@@ -120,35 +144,39 @@ bool insertToDb(const char *date, const char *title, const char *description, co
     return true;
 }
 
-bool updateToDb(const char *date, const char *title, const char *description, const char *start,
-                const char *duration,
-                int repeatCycle, const char *endDate, const char *finish, bool repeats[7],
-                int eventID,
-                const char *filepath) {
+bool updateToDb(const char *day, const char *month, const char *year, const char *title,
+                const char *description, const char *start, const char *duration, int repeatCycle,
+                const char *endDay, const char *endMonth, const char *endYear,
+                bool repeats[7],int eventID, const char *filepath) {
     sqlite3 *db;
     char *zErrMsg = 0;
     int rc;
-    bool result;
     sqlite3_stmt *stmt;
 
     /* Open database */
     rc = sqlite3_open(filepath, &db);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
 
     char *updateQuery = "UPDATE CALENDAR SET(endDate) VALUES(?) Where eventID = ?;";
     printf("Query: %s\n", updateQuery);
 
     rc = sqlite3_prepare_v2(db, updateQuery, -1, &stmt, 0);
     if (rc == SQLITE_OK) {
-        sqlite3_bind_text(stmt, 1, date, strlen(date), 0);
-        sqlite3_bind_text(stmt, 2, eventID, strlen(eventID), 0);
+        sqlite3_bind_text(stmt, 1, day, (int)strlen(day), 0);
+        sqlite3_bind_text(stmt, 2, month, (int)strlen(month), 0);
+        sqlite3_bind_text(stmt, 3, year,(int) strlen(year), 0);
+        sqlite3_bind_int(stmt, 4, eventID);
     } else {
         fprintf(stderr, "Failed to execute statement: %s\n", sqlite3_errmsg(db));
     }
     sqlite3_step(stmt);
     sqlite3_finalize(stmt);
     sqlite3_close(db);
-    return insertToDb(date, title, description, start, duration, repeatCycle, endDate, repeats,
-                      filepath);
+    return insertToDb(day,month,year, title, description, start, duration, repeatCycle,
+                      endDay,endMonth,endYear, repeats,filepath);
 }
 
 bool deleteFromDb(int eventID, const char *filepath) {
@@ -159,6 +187,10 @@ bool deleteFromDb(int eventID, const char *filepath) {
 
     /* Open database */
     rc = sqlite3_open(filepath, &db);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
 
     char *deleteQuery = "DELETE FROM CALENDAR WHERE eventID = ?;";
     printf("Queury: %s\n", deleteQuery);
@@ -166,7 +198,7 @@ bool deleteFromDb(int eventID, const char *filepath) {
     rc = sqlite3_prepare_v2(db, deleteQuery, -1, &stmt, 0);
 
     if (rc == SQLITE_OK) {
-        sqlite3_bind_text(stmt, 1, eventID, strlen(eventID), 0);
+        sqlite3_bind_int(stmt, 1, eventID);
     } else {
         fprintf(stderr, "Failed to execute statement: %s\n", sqlite3_errmsg(db));
     }
@@ -186,12 +218,10 @@ bool selectFromDB(const char *start, const char *end, bool repeats[7], const cha
 
     /* Open database */
     rc = sqlite3_open(filepath, &db);
-//    string selectedDate = nativeDate;
-//    string delimiter = "/";
-//    string day = selectedDate.substr(0, selectedDate.find(delimiter));
-//    string rest = selectedDate.substr(selectedDate.find(delimiter) + 1);
-//    string month = rest.substr(0, selectedDate.find(delimiter));
-//    string year = rest.substr(selectedDate.find(delimiter) + 1);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
 
 
     const char *selectQuery = "select * from CALENDAR "
@@ -209,11 +239,11 @@ bool selectFromDB(const char *start, const char *end, bool repeats[7], const cha
 
     rc = sqlite3_prepare_v2(db, selectQuery, -1, &stmt, 0);
     if (rc == SQLITE_OK) {
-        int binddate = sqlite3_bind_text(stmt, 1, start, strlen(start), 0);
-        int bindenddate = sqlite3_bind_text(stmt, 2, end, strlen(end), 0);;
+        sqlite3_bind_text(stmt, 1, start, (int)strlen(start), 0);
+        sqlite3_bind_text(stmt, 2, end, (int)strlen(end), 0);;
         int i;
         for (i = 0; i < 7; i++) {
-            int days = sqlite3_bind_int(stmt, i + 3, repeats[i]);
+            sqlite3_bind_int(stmt, i + 3, repeats[i]);
         }
     } else {
         fprintf(stderr, "Failed to execute statement: %s\n", sqlite3_errmsg(db));
@@ -249,6 +279,10 @@ bool displayDb(const char *filepath) {
 
     /* Open database */
     rc = sqlite3_open(filepath, &db);
+    if (rc != SQLITE_OK) {
+        sqlite3_free(zErrMsg);
+        return false;
+    }
 
     const char *selectQuery = "select * from calendar join weeklyID on calendar.weeklyID = weekly.weeklyID;";
     printf("Queury: %s\n\n", selectQuery);
